+++ docs/source/releases.rst
@@ -10,6 +10,7 @@ Fixes
 =====
 
 * Address problem where critical failures are missed in testing (`#262 <https://github.com/Electrostatics/pdb2pqr/issues/262>`_).  This doesn't represent a change to the (documented) API but will change functionality for external code that isn't ready to handle the raised error.
+* Fixes calculation of non-integer charge (`#264 <https://github.com/Electrostatics/pdb2pqr/issues/264>_`).
 
 Changes
 =======